Move collection to /; delete /search, /about, CLIR sub-nav
URL changes:
- /collection content moved to / (root)
- /search and /about deleted (404)
- /collection deleted (404; not live so no dead-link concern)

Layout: artwork/ephemera cohort indicator now sits on the same row
as the CGPA ARCHIVE title (right-aligned via flex justify-between)
instead of stacking below it.

Header simplified: removed the entire CLIR sub-nav (Collection /
Browse / Artists / Search / About). The Creative Growth white-label
nav at the top stays. Search button wiring on the CG header is also
removed since /search no longer exists.

Updated all internal links accordingly:
- src/app/page.tsx now hosts the collection (was the landing page)
- CohortNav.tsx: Artwork link points at / instead of /collection
- artwork detail breadcrumb + category links point to / with
  ?theme= or ?format= depending on category kind
- Footer + 404 pages no longer link to /collection or /about

Dead components removed: CategoryTabs.tsx (was used by old /collection
which used FilterBar instead) and SearchBar.tsx (was used only by
deleted /search page).
