+++ src/app/page.tsx
@@ -1,100 +1,104 @@
-import Link from "next/link";
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { createServerSupabaseClient } from "@/lib/supabase/server";
 import ArtworkGrid from "@/components/ArtworkGrid";
 import ArtworkCard from "@/components/ArtworkCard";
-import { createServerSupabaseClient } from "@/lib/supabase/server";
-import { Artwork } from "@/lib/types";
+import Pagination from "@/components/Pagination";
+import FilterBar from "@/components/FilterBar";
+import CohortNav from "@/components/CohortNav";
+import { parseSearchParams } from "@/lib/filter-state";
+import { queryArtworks, getFacetCounts } from "@/lib/collection-query";
 
-async function getFeaturedArtworks(): Promise<
-  (Artwork & { artist?: { id: string; first_name: string; last_name: string } })[]
-> {
-  const supabase = createServerSupabaseClient();
+const ITEMS_PER_PAGE = 24;
 
-  const { data, error } = await supabase
-    .from("artworks")
-    .select(
-      `
-      *,
-      artist:artists(id, first_name, last_name)
-      `
-    )
-    .eq("on_website", true)
-    .order("sort_order", { ascending: true })
-    .limit(12);
-
-  if (error) {
-    console.error("Error fetching artworks:", error);
-    return [];
-  }
-
-  return data || [];
+interface HomePageProps {
+  searchParams: Promise<Record<string, string | string[] | undefined>>;
 }
 
-export default async function Home() {
-  const artworks = await getFeaturedArtworks();
+export const metadata = {
+  title: "CGPA Archive | Creative Growth Gallery",
+  description: "Browse the complete collection of artworks",
+};
+
+export default async function HomePage({ searchParams }: HomePageProps) {
+  const raw = await searchParams;
+  const state = parseSearchParams(raw);
+  const supabase = createServerSupabaseClient();
+
+  const [{ artworks, total }, facets, formatCats, themeCats, allArtists] = await Promise.all([
+    queryArtworks(supabase, state, "artwork"),
+    getFacetCounts(supabase, state, "artwork"),
+    supabase.from("categories").select("name, slug").eq("kind", "format").order("name"),
+    supabase.from("categories").select("name, slug").eq("kind", "theme").order("name"),
+    supabase.from("artists").select("slug, first_name, last_name").order("last_name").order("first_name"),
+  ]);
+
+  const themeOptions = (themeCats.data || []).map((c) => ({
+    value: c.slug,
+    label: c.name,
+    count: facets.themes[c.slug] || 0,
+  }));
+  const formatOptions = (formatCats.data || []).map((c) => ({
+    value: c.slug,
+    label: c.name,
+    count: facets.formats[c.slug] || 0,
+  }));
+  const decadeOptions = Object.keys(facets.decades).sort().map((d) => ({
+    value: d,
+    label: d,
+    count: facets.decades[d],
+  }));
+  const artistOptions = (allArtists.data || []).map((a) => ({
+    slug: a.slug,
+    name: `${a.first_name} ${a.last_name}`.trim(),
+    available: facets.availableArtistSlugs.has(a.slug),
+  }));
+
+  const totalPages = Math.ceil(total / ITEMS_PER_PAGE);
 
   return (
-    <div>
-      {/* Hero Section */}
-      <section className="bg-gray-50 border-b border-gray-200">
-        <div className="container-max py-16 md:py-24">
-          <div className="max-w-2xl">
-            <h1 className="font-serif text-4xl md:text-5xl font-bold text-gray-900 mb-4">
-              Creative Growth Art Collection
-            </h1>
-            <p className="text-lg text-gray-600 mb-8">
-              Explore a carefully curated digital collection of contemporary art
-              from the Creative Growth Art Center, digitized with support from
-              the Council on Library and Information Resources.
-            </p>
-            <div className="flex gap-4">
-              <Link
-                href="/collection"
-                className="button-primary"
-              >
-                Browse Collection
-              </Link>
-              <Link
-                href="/artists"
-                className="button-secondary"
-              >
-                View Artists
-              </Link>
-            </div>
-          </div>
+    <div className="container-max py-12">
+      <div className="flex items-baseline justify-between gap-6 mb-6 flex-wrap">
+        <h1 className="font-sans text-5xl font-bold text-gray-900 tracking-tight">CGPA ARCHIVE</h1>
+        <CohortNav active="artwork" />
+      </div>
+
+      <FilterBar
+        state={state}
+        cohort="artwork"
+        themeOptions={themeOptions}
+        formatOptions={formatOptions}
+        decadeOptions={decadeOptions}
+        artistOptions={artistOptions}
+      />
+
+      {artworks.length > 0 ? (
+        <>
+          <p className="text-gray-600 mb-6 text-sm">
+            {total} {total === 1 ? "work" : "works"}
+            {state.q && <> for &ldquo;{state.q}&rdquo;</>}
+          </p>
+
+          <ArtworkGrid>
+            {artworks.map((artwork) => (
+              <ArtworkCard key={artwork.id} artwork={artwork as any} />
+            ))}
+          </ArtworkGrid>
+
+          {totalPages > 1 && (
+            <Pagination
+              currentPage={state.page}
+              totalPages={totalPages}
+              baseUrl="/"
+              preserveParams={["q", "theme", "format", "decade", "artist", "sort"]}
+            />
+          )}
+        </>
+      ) : (
+        <div className="text-center py-12">
+          <p className="text-gray-600 text-lg mb-4">No artworks match your filters.</p>
+          <a href="/" className="text-blue-600 underline">Clear filters</a>
         </div>
-      </section>
-
-      {/* Featured Artworks */}
-      <section className="container-max py-16">
-        <h2 className="font-serif text-3xl font-bold text-gray-900 mb-8">
-          Featured Works
-        </h2>
-
-        {artworks.length > 0 ? (
-          <>
-            <ArtworkGrid>
-              {artworks.map((artwork) => (
-                <ArtworkCard key={artwork.id} artwork={artwork} />
-              ))}
-            </ArtworkGrid>
-
-            <div className="mt-12 text-center">
-              <Link
-                href="/collection"
-                className="text-blue-600 hover:text-blue-800 font-medium text-lg"
-              >
-                View all artworks →
-              </Link>
-            </div>
-          </>
-        ) : (
-          <div className="text-center py-12">
-            <p className="text-gray-600">
-              No artworks available yet. Check back soon.
-            </p>
-          </div>
-        )}
-      </section>
+      )}
     </div>
   );
 }
