--- src/components/SearchBar.tsx
@@ -1,48 +1,0 @@
-"use client";
-
-import { FormEvent, useState } from "react";
-import { useRouter } from "next/navigation";
-
-export default function SearchBar() {
-  const [query, setQuery] = useState("");
-  const router = useRouter();
-
-  const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (query.trim()) {
-      router.push(`/search?q=${encodeURIComponent(query)}`);
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="flex items-center gap-2">
-      <input
-        type="search"
-        placeholder="Search artworks..."
-        value={query}
-        onChange={(e) => setQuery(e.target.value)}
-        aria-label="Search artworks and artists"
-        className="px-3 py-2 border border-gray-300 rounded focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-      />
-      <button
-        type="submit"
-        className="p-2 text-gray-700 hover:text-blue-600 transition-colors"
-        aria-label="Search"
-      >
-        <svg
-          className="w-5 h-5"
-          fill="none"
-          stroke="currentColor"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-          />
-        </svg>
-      </button>
-    </form>
-  );
-}
